requirements.txt
asgiref==3.5.2
Django==4.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
Pillow==9.2.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
pytz==2022.2.1
sqlparse==0.4.2
whitenoise==6.2.0
